requirements.txt
djangorestframework==3.8.2
django==2.1.1
markdown==2.6.11
django-filter==2.0.0
uwsgi==2.0.17.1
psycopg2-binary==2.7.5
django-oauth-toolkit==1.2.0
urllib3==1.23
numpy==1.16
django-hamlpy==1.1.1
bokeh==2.0.2
tornado==5.1.1
pandas<0.26
holoviews==1.13.2
recordclass==0.5
geoviews==1.8.1
cartopy==0.17.0
pillow==5.2.0
scipy==1.1.0
IPython==7.0.1
pytz==2018.5
python-dateutil==2.7.3
panel==0.9.0
param==1.9.3
jupyter<7.0.0
ipykernel<6.0.0
gunicorn==19.9.0
geopandas<0.8
scikit-learn<0.20
feather-format==0.3.1
pyarrow==0.12.0
jupyter-core
drf-yasg
django-prettyjson
matplotlib<3.0.0
datashader
xarray
pscript
ttn
paho-mqtt
shapely==1.7.0
methodtools
